rebane CarrotCoreBase2 to CarrotCoreBase
