+++ carrot-core/src/test/java/org/bigbase/carrot/BigSortedMapLargeKVsTest.java
@@ -31,7 +31,7 @@ import org.bigbase.carrot.util.Utils;
 import org.junit.Before;
 import org.junit.Test;
 
-public class BigSortedMapLargeKVsTest extends CarrotCoreBase2 {
+public class BigSortedMapLargeKVsTest extends CarrotCoreBase {
 
   private static final Logger log = LogManager.getLogger(BigSortedMapLargeKVsTest.class);
