+++ carrot-core/src/test/java/org/bigbase/carrot/redis/hashes/HashesAPITest.java
@@ -22,13 +22,13 @@ import java.util.*;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.bigbase.carrot.BigSortedMap;
-import org.bigbase.carrot.CarrotCoreBase2;
+import org.bigbase.carrot.CarrotCoreBase;
 import org.bigbase.carrot.ops.OperationFailedException;
 import org.bigbase.carrot.util.Pair;
 import org.bigbase.carrot.util.Utils;
 import org.junit.Test;
 
-public class HashesAPITest extends CarrotCoreBase2 {
+public class HashesAPITest extends CarrotCoreBase {
 
   private static final Logger log = LogManager.getLogger(HashesAPITest.class);
