+++ carrot-core/src/test/java/org/bigbase/carrot/redis/lists/ListsTest.java
@@ -23,7 +23,7 @@ import java.util.Random;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.bigbase.carrot.BigSortedMap;
-import org.bigbase.carrot.CarrotCoreBase2;
+import org.bigbase.carrot.CarrotCoreBase;
 import org.bigbase.carrot.DataBlock;
 import org.bigbase.carrot.redis.lists.Lists.Side;
 import org.bigbase.carrot.util.Key;
@@ -34,7 +34,7 @@ import org.junit.Before;
 import org.junit.Ignore;
 import org.junit.Test;
 
-public class ListsTest extends CarrotCoreBase2 {
+public class ListsTest extends CarrotCoreBase {
 
   private static final Logger log = LogManager.getLogger(ListsTest.class);
