+++ carrot-core/src/test/java/org/bigbase/carrot/redis/sets/SetScannerTest.java
@@ -21,7 +21,7 @@ import java.util.Random;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.bigbase.carrot.BigSortedMap;
-import org.bigbase.carrot.CarrotCoreBase2;
+import org.bigbase.carrot.CarrotCoreBase;
 import org.bigbase.carrot.util.Key;
 import org.bigbase.carrot.util.UnsafeAccess;
 import org.bigbase.carrot.util.Utils;
@@ -30,7 +30,7 @@ import org.junit.Test;
 
 import static org.junit.Assert.*;
 
-public class SetScannerTest extends CarrotCoreBase2 {
+public class SetScannerTest extends CarrotCoreBase {
 
   private static final Logger log = LogManager.getLogger(SetScannerTest.class);
