+++ carrot-core/src/test/java/org/bigbase/carrot/redis/sets/SetsAPITest.java
@@ -19,7 +19,7 @@ import java.util.*;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.bigbase.carrot.BigSortedMap;
-import org.bigbase.carrot.CarrotCoreBase2;
+import org.bigbase.carrot.CarrotCoreBase;
 import org.bigbase.carrot.util.UnsafeAccess;
 import org.bigbase.carrot.util.Utils;
 import org.junit.Before;
@@ -27,7 +27,7 @@ import org.junit.Test;
 
 import static org.junit.Assert.*;
 
-public class SetsAPITest extends CarrotCoreBase2 {
+public class SetsAPITest extends CarrotCoreBase {
 
   private static final Logger log = LogManager.getLogger(SetsAPITest.class);
