+++ carrot-core/src/test/java/org/bigbase/carrot/redis/sets/SetsTest.java
@@ -25,7 +25,7 @@ import java.util.zip.GZIPOutputStream;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.bigbase.carrot.BigSortedMap;
-import org.bigbase.carrot.CarrotCoreBase2;
+import org.bigbase.carrot.CarrotCoreBase;
 import org.bigbase.carrot.OrphanMemoryStats;
 import org.bigbase.carrot.compression.Codec;
 import org.bigbase.carrot.compression.CodecFactory;
@@ -41,7 +41,7 @@ import org.junit.Test;
 
 import static org.junit.Assert.*;
 
-public class SetsTest extends CarrotCoreBase2 {
+public class SetsTest extends CarrotCoreBase {
 
   private static final Logger log = LogManager.getLogger(SetsTest.class);
