+++ carrot-core/src/test/java/org/bigbase/carrot/redis/sparse/SparseBitmapsTest.java
@@ -23,7 +23,7 @@ import java.util.*;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.bigbase.carrot.BigSortedMap;
-import org.bigbase.carrot.CarrotCoreBase2;
+import org.bigbase.carrot.CarrotCoreBase;
 import org.bigbase.carrot.OrphanMemoryStats;
 import org.bigbase.carrot.redis.util.Commons;
 import org.bigbase.carrot.util.Key;
@@ -32,7 +32,7 @@ import org.bigbase.carrot.util.Utils;
 import org.junit.Before;
 import org.junit.Test;
 
-public class SparseBitmapsTest extends CarrotCoreBase2 {
+public class SparseBitmapsTest extends CarrotCoreBase {
 
   private static final Logger log = LogManager.getLogger(SparseBitmapsTest.class);
