+++ carrot-core/src/test/java/org/bigbase/carrot/redis/zsets/ZSetsAPITest.java
@@ -24,13 +24,13 @@ import java.util.*;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.bigbase.carrot.BigSortedMap;
-import org.bigbase.carrot.CarrotCoreBase2;
+import org.bigbase.carrot.CarrotCoreBase;
 import org.bigbase.carrot.util.Pair;
 import org.bigbase.carrot.util.Utils;
 import org.junit.Ignore;
 import org.junit.Test;
 
-public class ZSetsAPITest extends CarrotCoreBase2 {
+public class ZSetsAPITest extends CarrotCoreBase {
 
   private static final Logger log = LogManager.getLogger(ZSetsAPITest.class);
